update code - still need some work
