+++ conf/data/segmentation/kitti_img.yaml
@@ -1,0 +1,20 @@
+data:
+  task: segmentation
+  class: kitti.S3DIS1x1Dataset
+  dataroot: data
+  train_transforms:
+    - transform: SphericalProjection
+      params:
+        img_H: ''
+        img_W: ''
+        fov_UP: -3
+        fov_DOWN: 25
+        normalise: True
+    - transform: AddFeatsByKeys
+      params:
+        list_add_to_x: [True, True]
+        feat_names: ['', '']
+        delete_feats: [True, True]
+
+  test_transforms: ${data.train_transforms}
+  val_transforms: ${data.train_transforms}
