+++ test/test_transform.py
@@ -291,12 +291,12 @@ class Testhelpers(unittest.TestCase):
         torch.testing.assert_allclose(d.pos, torch.tensor([[2, 0, 0], [0, 2, 2]]).float())
 
     def test_normalizeFeature(self):
-        tr = NormalizeFeature(feature_name='new_feature', standardize=False)
-        d = Data(new_feature=torch.tensor([[-3,-2,-1,0,1,2]]).float())
+        tr = NormalizeFeature(feature_name="new_feature", standardize=False)
+        d = Data(new_feature=torch.tensor([[-3, -2, -1, 0, 1, 2]]).float())
         d = tr(d)
         torch.testing.assert_allclose(d.new_feature, torch.tensor([[0, 0.2, 0.4, 0.6, 0.8, 1]]).float())
 
-        tr = NormalizeFeature(feature_name='new_feature', standardize=True)
+        tr = NormalizeFeature(feature_name="new_feature", standardize=True)
         d = Data(new_feature=torch.tensor([[0, 1]]).float())
         d = tr(d)
         torch.testing.assert_allclose(d.new_feature, (d.new_feature - d.new_feature.mean()) / d.new_feature.std())
