+++ torch_points3d/core/data_transform/__init__.py
@@ -9,6 +9,7 @@ from .feature_augment import *
 from .features import *
 from .filters import *
 from .precollate import *
+from .projections import *
 
 _custom_transforms = sys.modules[__name__]
 _torch_geometric_transforms = sys.modules["torch_geometric.transforms"]
