+++ torch_points3d/modules/PPNet/blocks.py
@@ -15,23 +15,24 @@ class SimpleBlock(BaseModule):
     simple layer with PosPool
     we can perform a stride version (just change the query and the neighbors)
     """
+
     CONV_TYPE = ConvolutionFormat.PARTIAL_DENSE.value
     DENSITY_PARAMETER = 2.5
 
     def __init__(
-            self,
-            down_conv_nn=None,
-            grid_size=None,
-            prev_grid_size=None,
-            sigma=1.0,
-            max_num_neighbors=16,
-            position_embedding='xyz',
-            reduction='avg',
-            output_conv=False,
-            activation=torch.nn.LeakyReLU(negative_slope=0.2),
-            bn_momentum=0.01,
-            bn=FastBatchNorm1d,
-            **kwargs,
+        self,
+        down_conv_nn=None,
+        grid_size=None,
+        prev_grid_size=None,
+        sigma=1.0,
+        max_num_neighbors=16,
+        position_embedding="xyz",
+        reduction="avg",
+        output_conv=False,
+        activation=torch.nn.LeakyReLU(negative_slope=0.2),
+        bn_momentum=0.01,
+        bn=FastBatchNorm1d,
+        **kwargs,
     ):
         super(SimpleBlock, self).__init__()
         assert len(down_conv_nn) == 2
@@ -40,15 +41,17 @@ class SimpleBlock(BaseModule):
         search_radius = self.DENSITY_PARAMETER * sigma * prev_grid_size
         self.neighbour_finder = RadiusNeighbourFinder(search_radius, max_num_neighbors, conv_type=self.CONV_TYPE)
 
-        self.pospool = PosPoolLayer(num_inputs,
-                                    num_outputs,
-                                    search_radius,
-                                    position_embedding=position_embedding,
-                                    reduction=reduction,
-                                    output_conv=output_conv,
-                                    activation=activation,
-                                    bn_momentum=bn_momentum,
-                                    bn=bn)
+        self.pospool = PosPoolLayer(
+            num_inputs,
+            num_outputs,
+            search_radius,
+            position_embedding=position_embedding,
+            reduction=reduction,
+            output_conv=output_conv,
+            activation=activation,
+            bn_momentum=bn_momentum,
+            bn=bn,
+        )
 
         is_strided = prev_grid_size != grid_size
         if is_strided:
@@ -91,23 +94,24 @@ class SimpleInputBlock(BaseModule):
     a 1x1 conv and a simple layer with PosPool for input data
     we can perform a stride version (just change the query and the neighbors)
     """
+
     CONV_TYPE = ConvolutionFormat.PARTIAL_DENSE.value
     DENSITY_PARAMETER = 2.5
 
     def __init__(
-            self,
-            down_conv_nn=None,
-            grid_size=None,
-            prev_grid_size=None,
-            sigma=1.0,
-            max_num_neighbors=16,
-            position_embedding='xyz',
-            reduction='avg',
-            output_conv=False,
-            activation=torch.nn.LeakyReLU(negative_slope=0.2),
-            bn_momentum=0.01,
-            bn=FastBatchNorm1d,
-            **kwargs,
+        self,
+        down_conv_nn=None,
+        grid_size=None,
+        prev_grid_size=None,
+        sigma=1.0,
+        max_num_neighbors=16,
+        position_embedding="xyz",
+        reduction="avg",
+        output_conv=False,
+        activation=torch.nn.LeakyReLU(negative_slope=0.2),
+        bn_momentum=0.01,
+        bn=FastBatchNorm1d,
+        **kwargs,
     ):
         super(SimpleInputBlock, self).__init__()
         assert len(down_conv_nn) == 3
@@ -123,15 +127,17 @@ class SimpleInputBlock(BaseModule):
         search_radius = self.DENSITY_PARAMETER * sigma * prev_grid_size
         self.neighbour_finder = RadiusNeighbourFinder(search_radius, max_num_neighbors, conv_type=self.CONV_TYPE)
 
-        self.pospool = PosPoolLayer(d_2,
-                                    num_outputs,
-                                    search_radius,
-                                    position_embedding=position_embedding,
-                                    reduction=reduction,
-                                    output_conv=output_conv,
-                                    activation=activation,
-                                    bn_momentum=bn_momentum,
-                                    bn=bn)
+        self.pospool = PosPoolLayer(
+            d_2,
+            num_outputs,
+            search_radius,
+            position_embedding=position_embedding,
+            reduction=reduction,
+            output_conv=output_conv,
+            activation=activation,
+            bn_momentum=bn_momentum,
+            bn=bn,
+        )
 
         is_strided = prev_grid_size != grid_size
         if is_strided:
@@ -191,21 +197,21 @@ class ResnetBBlock(BaseModule):
     CONV_TYPE = ConvolutionFormat.PARTIAL_DENSE.value
 
     def __init__(
-            self,
-            down_conv_nn=None,
-            grid_size=None,
-            prev_grid_size=None,
-            sigma=1,
-            max_num_neighbors=16,
-            position_embedding='xyz',
-            reduction='avg',
-            output_conv=False,
-            activation=torch.nn.LeakyReLU(negative_slope=0.2),
-            has_bottleneck=True,
-            bottleneck_ratio=2,
-            bn_momentum=0.01,
-            bn=FastBatchNorm1d,
-            **kwargs,
+        self,
+        down_conv_nn=None,
+        grid_size=None,
+        prev_grid_size=None,
+        sigma=1,
+        max_num_neighbors=16,
+        position_embedding="xyz",
+        reduction="avg",
+        output_conv=False,
+        activation=torch.nn.LeakyReLU(negative_slope=0.2),
+        has_bottleneck=True,
+        bottleneck_ratio=2,
+        bn_momentum=0.01,
+        bn=FastBatchNorm1d,
+        **kwargs,
     ):
         super(ResnetBBlock, self).__init__()
         assert len(down_conv_nn) == 2, "down_conv_nn should be of size 2"
@@ -232,7 +238,8 @@ class ResnetBBlock(BaseModule):
             output_conv=output_conv,
             activation=activation,
             bn_momentum=bn_momentum,
-            bn=bn)
+            bn=bn,
+        )
 
         if self.has_bottleneck:
             if bn:
@@ -316,19 +323,19 @@ class PPStageBlock(BaseModule):
     """
 
     def __init__(
-            self,
-            block_names=None,
-            down_conv_nn=None,
-            grid_size=None,
-            prev_grid_size=None,
-            has_bottleneck=None,
-            bottleneck_ratio=None,
-            max_num_neighbors=None,
-            position_embedding=None,
-            reduction=None,
-            output_conv=None,
-            bn_momentum=None,
-            **kwargs,
+        self,
+        block_names=None,
+        down_conv_nn=None,
+        grid_size=None,
+        prev_grid_size=None,
+        has_bottleneck=None,
+        bottleneck_ratio=None,
+        max_num_neighbors=None,
+        position_embedding=None,
+        reduction=None,
+        output_conv=None,
+        bn_momentum=None,
+        **kwargs,
     ):
         super(PPStageBlock, self).__init__()
